Update auth and shows layouts
