+++ tv-shows/src/app/(shows)/all-shows/[id]/page.tsx
@@ -1,0 +1,5 @@
+import ShowContainer from '../../../../components/features/shows/ShowContainer/ShowContainer';
+
+export default function ShowDetailsPage() {
+	return <ShowContainer />;
+}
