+++ tv-shows/src/app/(shows)/all-shows/page.tsx
@@ -1,0 +1,5 @@
+import { AllShowsContainer } from '../../../components/features/shows/AllShowsContainer/AllShowsContainer';
+
+export default function AllShowsPage() {
+	return <AllShowsContainer />;
+}
