+++ tv-shows/src/app/(shows)/top-rated/page.tsx
@@ -1,0 +1,5 @@
+import { TopRatedShowsContainer } from '../../../components/features/shows/TopRatedShowsContainer.tsx/TopRatedShowsContainer';
+
+export default function TopRatedShowsPage() {
+	return <TopRatedShowsContainer />;
+}
