--- tv-shows/src/app/all-shows/[id]/page.tsx
@@ -1,11 +1,0 @@
-import ShowContainer from '../../../components/features/shows/ShowContainer/ShowContainer';
-import { AuthRedirect } from '../../../components/shared/AuthRedirect/AuthRedirect';
-
-export default function ShowDetailsPage() {
-	return (
-		<>
-			<AuthRedirect />
-			<ShowContainer />;
-		</>
-	);
-}
