--- tv-shows/src/app/all-shows/page.tsx
@@ -1,11 +1,0 @@
-import { AllShowsContainer } from '../../components/features/shows/AllShowsContainer/AllShowsContainer';
-import { AuthRedirect } from '../../components/shared/AuthRedirect/AuthRedirect';
-
-export default function AllShowsPage() {
-	return (
-		<>
-			<AuthRedirect />
-			<AllShowsContainer />;
-		</>
-	);
-}
