+++ tv-shows/src/app/auth/layout.tsx
@@ -1,4 +1,5 @@
 import { Flex } from '@chakra-ui/react';
+import { AuthRedirect } from '../../components/shared/AuthRedirect/AuthRedirect';
 
 export default function AuthLayout({
 	children,
@@ -6,8 +7,11 @@ export default function AuthLayout({
 	children: React.ReactNode;
 }>) {
 	return (
-		<Flex align="center" justify="center" h="100%">
-			{children}
-		</Flex>
+		<>
+			<AuthRedirect invert={true} />
+			<Flex as="main" flexGrow="1" align="center" justify="center" h="100%">
+				{children}
+			</Flex>
+		</>
 	);
 }
