+++ tv-shows/src/app/auth/register/page.tsx
@@ -1,11 +1,5 @@
 import { RegisterForm } from '../../../components/features/auth/RegisterForm/RegisterForm';
-import { AuthRedirect } from '../../../components/shared/AuthRedirect/AuthRedirect';
 
 export default function Register() {
-	return (
-		<>
-			<AuthRedirect invert={true} />
-			<RegisterForm />;
-		</>
-	);
+	return <RegisterForm />;
 }
