--- tv-shows/src/app/top-rated/page.tsx
@@ -1,11 +1,0 @@
-import { TopRatedShowsContainer } from '../../components/features/shows/TopRatedShowsContainer.tsx/TopRatedShowsContainer';
-import { AuthRedirect } from '../../components/shared/AuthRedirect/AuthRedirect';
-
-export default function TopRatedShowsPage() {
-	return (
-		<>
-			<AuthRedirect />;
-			<TopRatedShowsContainer />;
-		</>
-	);
-}
